[refactor/#63] my-reservations 페이지 컴포넌트 구조 개선
