+++ src/app/(header)/my-reservations/page.tsx
@@ -1,5 +1,5 @@
 import { SideNavigationMenu } from "@/components/layout/SideNavigationMenu";
-import { ReservationItem } from "@/components/ReservationHistory/ReservationItem";
+import { ReservationItem } from "@/components/my-reservations/ReservationItem";
 import { Dropdown, MenuItem } from "@/components/ui/Dropdown";
 import EmptyList from "@/assets/svgs/empty_list.svg";
 import Image from "next/image";
